Routine Organization/Refactiring
Cleaning up the repository... made comments in the index.html file and added a new file called node_modules/.package-lock.json
Attempts to template the project and make it more organized failed... ejs is not supported in TRULY static sites.
AWS S3 is not a good option for hosting static sites... it is not free and it is not easy to use.
AWS Amplify is a better option for hosting static sites... it is free and it is easy to use.
I will try to use Amplify as it will allow me to host a site for free.
This comes with a cost... I will have to have cluttered code. I will have to have a lot of files in the repository.
I may have to keep all the information on one page...
While not ideal, this is the best option for me at this time.