Extensive introductory and transition changes
Made the introduction to a parallax instead of the splitscreen effect
Over the introductory parallax there is a text box
When hovered over the textbox, the parallax will blur, making the text a bit easier to read
The transition section is now a full paragraph that makes it easier to transition between the introduction parllax and the about me parallax
Considering scrapping the projects section in favor of making it a link or something to an actual projects page. Not fully committed but about 90% to pulling the trigger.