Begin creationg of Projects page
Started new branch yesterday (4/23/24) for the projects page branch.
Planning on getting some of this running before and the look of it nice
before moving on to other sections. This is also a good way to version
out the code, branch the project, and make it a bit more organized in
the testing and development environment. More to come on this